requirements.txt
flask==2.0.1
flask-cors==6.0.0
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
python-dotenv==0.19.0
requests==2.26.0
openai==1.84.0 